BACKEND/requirements.txt
Flask
requests
python-dotenv
huggingface_hub
transformers